[issue-244-refactor-network-handling-code] more cleanup
- moved CanDirectFetch and UpdateBlockAvailability into Network:: namespace (NetworkMessage .h/.cpp)
- removed redundant empty destructors on NetworkMessage subclasses
- Network:: methods laid out alphabetically in both .h and .cpp files
- optimized includes in main.cpp
