+++ src/main.h
@@ -723,6 +723,4 @@ std::set<CBlockIndex*> setDirtyBlockIndex;
 /** Number of peers from which we're downloading blocks. */
 int nPeersWithValidatedDownloads = 0;
 
-
-
 #endif // BITCOIN_MAIN_H
